+++ Project programming/Database/User.cs
@@ -8,10 +8,10 @@ namespace Classes
 {
     public class User
     {
-        public int Id{ get; set; }
+        public ushort Id { get; set; }
         public string Name { get; set; }
         public string Password { get; set; }
-        public int Salary { get; set; }
+        public uint Salary { get; set; }
         public string Email { get; set; }
         public User( string Name , string Password , string email) 
         {
